feat:added CartContextProvider without Reducer
