+++ src/App.js
@@ -10,6 +10,8 @@ import * as api from "./api";
 import useLocalStorage from "./hooks/useLocalStorage";
 import loadLocalStorageItems from "./utils/loadLocalStorageItems";
 
+import CartContextProvider from "./components/ContextComponents/CartContextProvider";
+
 // Checkout constants
 import {
   PROFILE,
@@ -24,36 +26,36 @@ import {
   NEWPROD_URL,
 } from "./utils/constants";
 
-function buildNewCartItem(cartItem) {
-  if (cartItem.quantity >= cartItem.unitsInStock) {
-    return cartItem;
-  }
+// function buildNewCartItem(cartItem) {
+//   if (cartItem.quantity >= cartItem.unitsInStock) {
+//     return cartItem;
+//   }
 
-  return {
-    id: cartItem.id,
-    title: cartItem.title,
-    img: cartItem.img,
-    price: cartItem.price,
-    unitsInStock: cartItem.unitsInStock,
-    createdAt: cartItem.createdAt,
-    updatedAt: cartItem.updatedAt,
-    quantity: cartItem.quantity + 1,
-  };
-}
+//   return {
+//     id: cartItem.id,
+//     title: cartItem.title,
+//     img: cartItem.img,
+//     price: cartItem.price,
+//     unitsInStock: cartItem.unitsInStock,
+//     createdAt: cartItem.createdAt,
+//     updatedAt: cartItem.updatedAt,
+//     quantity: cartItem.quantity + 1,
+//   };
+// }
 
 const PRODUCTS_LOCAL_STORAGE_KEY = "react-sc-state-products";
-const CART_ITEMS_LOCAL_STORAGE_KEY = "react-sc-state-cart-items";
+// const CART_ITEMS_LOCAL_STORAGE_KEY = "react-sc-state-cart-items";
 
 function App() {
   const [products, setProducts] = useState(() =>
     loadLocalStorageItems(PRODUCTS_LOCAL_STORAGE_KEY, []),
   );
-  const [cartItems, setCartItems] = useState(() =>
-    loadLocalStorageItems(CART_ITEMS_LOCAL_STORAGE_KEY, []),
-  );
+  // const [cartItems, setCartItems] = useState(() =>
+  //   loadLocalStorageItems(CART_ITEMS_LOCAL_STORAGE_KEY, []),
+  // );
 
   useLocalStorage(products, PRODUCTS_LOCAL_STORAGE_KEY);
-  useLocalStorage(cartItems, CART_ITEMS_LOCAL_STORAGE_KEY);
+  // useLocalStorage(cartItems, CART_ITEMS_LOCAL_STORAGE_KEY);
 
   const [isLoading, setIsLoading] = useState(false);
   const [hasError, setHasError] = useState(false);
@@ -77,54 +79,54 @@ function App() {
     }
   }, []);
 
-  function handleAddToCart(productId) {
-    const prevCartItem = cartItems.find((item) => item.id === productId);
-    const foundProduct = products.find((product) => product.id === productId);
+  // function handleAddToCart(productId) {
+  //   const prevCartItem = cartItems.find((item) => item.id === productId);
+  //   const foundProduct = products.find((product) => product.id === productId);
 
-    if (prevCartItem) {
-      const updatedCartItems = cartItems.map((item) => {
-        if (item.id !== productId) {
-          return item;
-        }
+  //   if (prevCartItem) {
+  //     const updatedCartItems = cartItems.map((item) => {
+  //       if (item.id !== productId) {
+  //         return item;
+  //       }
 
-        if (item.quantity >= item.unitsInStock) {
-          return item;
-        }
+  //       if (item.quantity >= item.unitsInStock) {
+  //         return item;
+  //       }
 
-        return {
-          ...item,
-          quantity: item.quantity + 1,
-        };
-      });
+  //       return {
+  //         ...item,
+  //         quantity: item.quantity + 1,
+  //       };
+  //     });
 
-      setCartItems(updatedCartItems);
-      return;
-    }
+  //     setCartItems(updatedCartItems);
+  //     return;
+  //   }
 
-    const updatedProduct = buildNewCartItem(foundProduct);
-    setCartItems((prevState) => [...prevState, updatedProduct]);
-  }
+  //   const updatedProduct = buildNewCartItem(foundProduct);
+  //   setCartItems((prevState) => [...prevState, updatedProduct]);
+  // }
 
-  function handleChange(event, productId) {
-    const updatedCartItems = cartItems.map((item) => {
-      if (item.id === productId && item.quantity <= item.unitsInStock) {
-        return {
-          ...item,
-          quantity: Number(event.target.value),
-        };
-      }
+  // function handleChange(event, productId) {
+  //   const updatedCartItems = cartItems.map((item) => {
+  //     if (item.id === productId && item.quantity <= item.unitsInStock) {
+  //       return {
+  //         ...item,
+  //         quantity: Number(event.target.value),
+  //       };
+  //     }
 
-      return item;
-    });
+  //     return item;
+  //   });
 
-    setCartItems(updatedCartItems);
-  }
+  //   setCartItems(updatedCartItems);
+  // }
 
-  function handleRemove(productId) {
-    const updatedCartItems = cartItems.filter((item) => item.id !== productId);
+  // function handleRemove(productId) {
+  //   const updatedCartItems = cartItems.filter((item) => item.id !== productId);
 
-    setCartItems(updatedCartItems);
-  }
+  //   setCartItems(updatedCartItems);
+  // }
 
   function handleDownVote(productId) {
     const updatedProducts = products.map((product) => {
@@ -195,61 +197,64 @@ function App() {
   }
 
   return (
-    <BrowserRouter>
-      <Switch>
-        <Route path={PROFILE_URL}>
-          <Checkout
-            processStep={PROFILE}
-            cartItems={cartItems}
-            handleRemove={handleRemove}
-            handleChange={handleChange}
-          />
-        </Route>
-        <Route path={BILLING_URL}>
-          <Checkout
-            processStep={BILLING}
-            cartItems={cartItems}
-            handleRemove={handleRemove}
-            handleChange={handleChange}
-          />
-        </Route>
-        <Route path={PAYMENT_URL}>
-          <Checkout
-            processStep={PAYMENT}
-            cartItems={cartItems}
-            handleRemove={handleRemove}
-            handleChange={handleChange}
-          />
-        </Route>
-        <Route path={SUMMARY_URL}>
-          <Checkout
-            processStep={SUMMARY}
-            cartItems={cartItems}
-            handleRemove={handleRemove}
-            handleChange={handleChange}
-          />
-        </Route>
-        <Route path={NEWPROD_URL}>
-          <NewProduct saveNewProduct={saveNewProduct} />
-        </Route>
-        <Route path={HOME_URL} exact>
-          <Home
-            fullWidth
-            cartItems={cartItems}
-            products={products}
-            isLoading={isLoading}
-            hasError={hasError}
-            loadingError={loadingError}
-            handleDownVote={handleDownVote}
-            handleUpVote={handleUpVote}
-            handleSetFavorite={handleSetFavorite}
-            handleAddToCart={handleAddToCart}
-            handleRemove={handleRemove}
-            handleChange={handleChange}
-          />
-        </Route>
-      </Switch>
-    </BrowserRouter>
+    <CartContextProvider>
+      <BrowserRouter>
+        {" "}
+        <Switch>
+          <Route path={PROFILE_URL}>
+            <Checkout
+              processStep={PROFILE}
+              // cartItems={cartItems}
+              // handleRemove={handleRemove}
+              // handleChange={handleChange}
+            />
+          </Route>
+          <Route path={BILLING_URL}>
+            <Checkout
+              processStep={BILLING}
+              // cartItems={cartItems}
+              // handleRemove={handleRemove}
+              // handleChange={handleChange}
+            />
+          </Route>
+          <Route path={PAYMENT_URL}>
+            <Checkout
+              processStep={PAYMENT}
+              // cartItems={cartItems}
+              // handleRemove={handleRemove}
+              // handleChange={handleChange}
+            />
+          </Route>
+          <Route path={SUMMARY_URL}>
+            <Checkout
+              processStep={SUMMARY}
+              // cartItems={cartItems}
+              // handleRemove={handleRemove}
+              // handleChange={handleChange}
+            />
+          </Route>
+          <Route path={NEWPROD_URL}>
+            <NewProduct saveNewProduct={saveNewProduct} />
+          </Route>
+          <Route path={HOME_URL} exact>
+            <Home
+              fullWidth
+              // cartItems={cartItems}
+              products={products}
+              isLoading={isLoading}
+              hasError={hasError}
+              loadingError={loadingError}
+              handleDownVote={handleDownVote}
+              handleUpVote={handleUpVote}
+              handleSetFavorite={handleSetFavorite}
+              // handleAddToCart={handleAddToCart}
+              // handleRemove={handleRemove}
+              // handleChange={handleChange}
+            />
+          </Route>
+        </Switch>
+      </BrowserRouter>
+    </CartContextProvider>
   );
 }
