+++ src/components/ContextComponents/CartContextProvider/CartContextProvider.js
@@ -1,0 +1,105 @@
+/* eslint-disable import/extensions */
+import React, { useState } from "react";
+
+import CartContext from "../../../context/cart-context";
+
+import loadLocalStorageItems from "../../../utils/loadLocalStorageItems";
+import useLocalStorage from "../../../hooks/useLocalStorage";
+
+const CART_ITEMS_LOCAL_STORAGE_KEY = "react-sc-state-cart-items";
+const PRODUCTS_LOCAL_STORAGE_KEY = "react-sc-state-products";
+
+function buildNewCartItem(cartItem) {
+  if (cartItem.quantity >= cartItem.unitsInStock) {
+    return cartItem;
+  }
+
+  return {
+    id: cartItem.id,
+    title: cartItem.title,
+    img: cartItem.img,
+    price: cartItem.price,
+    unitsInStock: cartItem.unitsInStock,
+    createdAt: cartItem.createdAt,
+    updatedAt: cartItem.updatedAt,
+    quantity: cartItem.quantity + 1,
+  };
+}
+
+function CartContextProvider({ children }) {
+  const [products] = useState(() =>
+    loadLocalStorageItems(PRODUCTS_LOCAL_STORAGE_KEY, []),
+  );
+  const [cartItems, setCartItems] = useState(() =>
+    loadLocalStorageItems(CART_ITEMS_LOCAL_STORAGE_KEY, []),
+  );
+
+  useLocalStorage(products, PRODUCTS_LOCAL_STORAGE_KEY);
+
+  useLocalStorage(cartItems, CART_ITEMS_LOCAL_STORAGE_KEY);
+
+  // Methods
+  function handleAddToCart(productId) {
+    const prevCartItem = cartItems.find((item) => item.id === productId);
+    const foundProduct = products.find((product) => product.id === productId);
+
+    if (prevCartItem) {
+      const updatedCartItems = cartItems.map((item) => {
+        if (item.id !== productId) {
+          return item;
+        }
+
+        if (item.quantity >= item.unitsInStock) {
+          return item;
+        }
+
+        return {
+          ...item,
+          quantity: item.quantity + 1,
+        };
+      });
+
+      setCartItems(updatedCartItems);
+      return;
+    }
+
+    const updatedProduct = buildNewCartItem(foundProduct);
+    setCartItems((prevState) => [...prevState, updatedProduct]);
+  }
+
+  function handleChange(event, productId) {
+    const updatedCartItems = cartItems.map((item) => {
+      if (item.id === productId && item.quantity <= item.unitsInStock) {
+        return {
+          ...item,
+          quantity: Number(event.target.value),
+        };
+      }
+
+      return item;
+    });
+
+    setCartItems(updatedCartItems);
+  }
+
+  function handleRemove(productId) {
+    const updatedCartItems = cartItems.filter((item) => item.id !== productId);
+
+    setCartItems(updatedCartItems);
+  }
+
+  return (
+    <CartContext.Provider
+      value={{
+        cartItems: cartItems,
+        add: handleAddToCart,
+        remove: handleRemove,
+        change: handleChange,
+      }}
+    >
+      {children}
+    </CartContext.Provider>
+  );
+}
+
+export default CartContextProvider;
