+++ src/components/ContextComponents/CartContextProvider/index.js
@@ -1,0 +1,1 @@
+export { default } from "./CartContextProvider";
