+++ src/context/cart-context/cart-context.js
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+
+const CartContext = createContext({
+  cartItems: [],
+  add: () => {},
+  remove: () => {},
+  change: () => {},
+});
+
+export default CartContext;
