+++ src/context/cart-context/index.js
@@ -1,0 +1,1 @@
+export { default } from "./cart-context";
